feat: Store Detail 관련 Rest API 생성
issue: #3

+++ src/main/java/com/kinopio/eatgo/store/dto/StoreResponseDto.java
@@ -1,12 +1,8 @@
 package com.kinopio.eatgo.store.dto;
 
-import java.sql.Date;
 import java.util.List;
 
-import com.kinopio.eatgo.store.entity.Menu;
 import com.kinopio.eatgo.store.entity.OpenInfo;
-import com.kinopio.eatgo.store.entity.Review;
-import com.kinopio.eatgo.store.entity.StoreHistory;
 import com.kinopio.eatgo.store.entity.Tag;
 
 import lombok.AllArgsConstructor;
@@ -40,7 +36,7 @@ public class StoreResponseDto {
 	private List<OpenInfo> openInfos;
 	
 	// 리뷰 평점 평균
-	private Integer ratingAverage;
+	private Float ratingAverage;
 	
 	
 }
